chore: organize note versions
